# Allow bypassing of serial in simulation