Previously we used to simulate the serial connection of our uart peripheral. By separating implementation and interface we can add a `uartSim` implementation that communicates in `BitVector 8` instead of serializing the data over UART which allows for great simulation speedups

Before:
```
Unittests
  Wishbone.CaptureUgn
    capture ugn self test:     OK (0.79s)
  Wishbone.DnaPortE2
    dna port self test:        OK (4.20s)
  Wishbone.Time
    time rust self test:       OK (13.51s)
  Wishbone.Axi
    axi stream rust self test: OK (21.76s)
    parseTestResults:          OK

All 5 tests passed (40.26s)
```
After:
```
 lucas@lonneker:~/bittide-hardware$ cabal run bittide-instances:unittests -- -p Wishbone
Unittests
  Wishbone.CaptureUgn
    capture ugn self test:     OK (0.51s)
  Wishbone.DnaPortE2
    dna port self test:        OK (3.11s)
  Wishbone.Time
    time rust self test:       OK (8.27s)
  Wishbone.Axi
    axi stream rust self test: OK (17.19s)
    parseTestResults:          OK

All 5 tests passed (29.10s)
```